backend/docling-service/requirements.txt
# AWS S3 interaction
boto3>=1.28.0
botocore>=1.31.0

# Environment variables management
python-dotenv>=1.0.0

# Docling libraries
docling>=1.0.0
docling-core>=1.0.0
docling-parse>=1.0.0

# FastAPI for API endpoints
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Additional utilities
python-multipart>=0.0.6  # For handling file uploads in FastAPI
pillow>=9.5.0  # For image processing